Progress within FitnessAppWrapper
Almost all functionality complete and tested for DietPlan, with the exception being the ability to edit. This means that one can now read and store the plans from a file, display all 7 days of plans, and write the current list of plans to the same file. Once the edit feature has been implemented work can begin on implementation of these same features for the ExercisePlan class.. AKA almost finished!